Added Migration Pre/Post Text Messages
Pre text warnings and/or messages can be specified in migration files. When running a migration with a pre text message the user will be redirected to a page displaying the message and prompted to either cancel the migration or continue with it. After running a batch of migrations any post text messages they had will be displayed alongside the migration name.