Beefed Up Post Text Message
Post text message will now show on the pretext message page if the migration is interrupted by  a pretext message. The post message now shows which migrations were run in the batch if it is interrupted by a pretext and also displays which weren't run and were cancelled by the message. Created a whole new layout removing all the header icons and stuff. It's just a blank layout with the styling as well as the $w->msg function. This was required to send the post text messages to this page.