feat: Insights PDF&template delivery (#161)
* feat: Export to PDF button

Creating the button and a sample template for Export to PDF

refs: 10244

* feat: Export to PDF action

Fetching insight in action for exporting to PDF and planning steps for action in comments

refs: 10244

* feat: Find template list

Adding service function to find a list of all available templates

refs: 10244

* feat: Filter template list

Filter list of template titles to only display those that include the insights name.

refs: 10244

* feat: Filter templates for insight

Get name of insight to filter templates available to choose from for exporting to PDF

refs: 10244

* feat: Filter with category

Use the category variable in the findTemplates template service function to filter templates shown to only show templates including the insight class name and PDF in the title

refs: 10244

* feat: unnecessary line

Removing unnecessary line and undefined, unnecessary variable.

refs: 10244

* feat: find corrct template

Changes to find correct template to display in list, and comments for more changes needed

refs: 10244

* feat: multi-col form drop-down

Changing drop-down into multi-col form and adding code to be edited for use in yet to be created POSt function

refs: 10244

* feat: GET and POST

making get and post functions for exporting to PDF

refs: 10244

* feat: Export to PDF popup

Changing export to PDF button from link to another page to a popup box

refs: 10244

* feat: send data, render, redirect

Sending data to the POST function, rendering the insight data using the template selected, and redirecting back to runInsight

refs: 10244

* feat: correct postUrl

Correct url in postUrl

refs: 10244

* feat: Update variables

Update the variables used in POST to get the values from the multiColForm $template_list

refs: 10244

* feat: preparing data

Creating and correcting variables to get correct data in the post function for rendering in the template and exporting to PDF

refs: 10244

* feat: Check exportpdf function

Checking how the exportpdf function that the report module uses works.

refs: 10244

* feat: export to pdf using run_data

Changing the exportpdf function to use the run_data variable instead of rows

refs: 10244

* feat: Using run_data

changing exportphp function to use run_data

refs: 10244

* feat: checking variables

Checking new variables for exportpdf function

refs: 10244

* feat: using run_data

Editing code to use run_data in exportpdf service function

refs: 10244

* feat: Template mandatory

Making selection of a template for exporting to a PDF mandatory

refs: 10244

* feat: template and data

Organising template and data for the table for exporting to PDF

refs: 10244]

* feat: Template required

Change where user is told choosing a template is required

refs: 10244

* feat: errors noted in VSCode

Fixing small errors in the export PDF action and service function that VSCode had flagged with red underlines

refs: 10244

* feat: Small changes

Checking things in export action

refs: 10244

* feat: working to fix table

Trying to get all data rows to display when exported to PDF

refs: 10244

* feat: Making template required

Changing form setup so that selecting a template is a requirement for export to PDF to work

refs: 10244

* feat: get data for PDF

Making sure the export to PDF function finds the table data and changing how the title will display when a template is used.

refs: 10244

* feat: Export PDF to template

Preparing to move pdf.php GET function into the runInsight template

refs: 10244

* feat: doc title for PDF

Changing how the title on the PDF document will appear when exported based on using a template or not using a template

refs: 10244

* feat: Notes on PDF export

Comments on things to be done to have GET function for pdf.php action to be in runInsight template

refs: 10244

* feat: re-create form

Recreating the form in the same style as the cmfive standard in the template. Buttons are now editable for closing. Parameters for select option yet to be moved.

refs: 10244

* feat: formatting

Fix spacing and other formatting errors noted by phpcs

refs: 10244

* feat: creating export PDF

creating button and new select for export to PDF function in insights

refs: 10244

* feat: comments in js

Comments on what needs to be placed in the java script

refs: 10244

* feat: variables and close

Creating variables for modal, button, and modal elements in javascript and adding function to display modal and make modal disappear

refs: 10244

* feat: Window close

Commenting out function to close modal when clicking out because behaving wrong. Editing to find correct functions that cmfive uses for this purpose

refs: 10244

* feat: Closing X

Moving closing x to correct position in modal and trying to get clicking on Export to PDF to grey out page. Trying to get x to close modal and allow clicking on page ability to resume for user
[body footer?]

refs: 10244

* feat: labels, close, and ajax

Add labels to modal including title select box and buttons. Fix issues with closing X, clicking outside modal, and cancel button. Setting up code ready to implement ajax with submit button to send form information to POST function.

refs: 10244

* feat: Set export to submit

Export button submits to export/pdf action

refs: 10244

* feat: Clean up

Removing excess white space and unnecessary comments (commented out code) in preparation for making a pull request

refs: 10244

* feat: remove console.log

remove console log lines which were added for the purpose of testing

refs: 10244

* feat: remove old code

Old commented out code removed

refs: 10244

* feat: jQuery to Vanilla

Updating jQuery code to Vanilla JavaScript

refs: 10244

* fix: check for arrays and objects (#143)

* fix: check for arrays and objects

added a check for object or arrays as array_key_esists is depreciated for use on objects causing warnings

refs: 10244

* feat: requested changes

Changes requested on GitHub to tread $fields as object and use the allow operator instead of the array operator

refs: 10244

* fix: requested changes

Changes requested on GitHub to tread $fields as object and use the allow operator instead of the array operator

refs: 10244

* fix: remove duplications

deleted duplicated lines

refs: 10244

* fix: check is fields object

add an elseif to check if the fields variable is an object

refs: 10244

Co-authored-by: Taliesin Millhouse <taliesinwrmillhouse@gmail.com>

* feat: Fixing foundation

Working to fix issue with foundation function

refs: 10151

* feat: converting to jQuery

Converting Vanilla JavaScript code to jQuery so that foundation function will work

refs: 10151

* feat: Remove select duplicate

Removing the duplicate of the select option appearing twice in the drop down wen choosing a template for PDF export
[body footer?]

refs: 10151

* feat: clean-up code

remove testing code and unnecessary comments

refs: 10151

* feat: Alt Approach On Insights Modal (#159)

* Rough cut proof, WIP

* Integral patches

* H-Scroll wide tables

* Fix default PDF heading layout

* Remove experimental code for simple implementation

* Confirmed use case for core_layout_tpl

* Portrait & Landscape on Insights pdf

* Removed some lint

Co-authored-by: D.Crannaford <derek@2pisoftware.com>

Co-authored-by: Alice <alice.hutley@gmail.com>
Co-authored-by: Taliesin Millhouse <taliesinwrmillhouse@gmail.com>
Co-authored-by: Alice Hutley <55948506+Tinystep1@users.noreply.github.com>
Co-authored-by: D.Crannaford <derek@2pisoftware.com>
Co-authored-by: Adam Buckley <adam.buckley90@gmail.com>