Fix bootstrap theme search
I do not like this solution:
Cmfive.openModal downloads the modal and appends it to the template modal box,
and any scripts inside the downloaded content are expected to execute.
However, as per MDN[1], setting element content with `innerHTML`
should not execute scripts as a security measure.
To get around this, I've had to manually `eval` any scripts the downloaded modal contains.

I've also added jquery to the bootstrap template as the search module uses it.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Element/innerHTML#security_considerations